# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Matka Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /markets/{id}/auto-config:
    put:
      operationId: updateMarketAutoConfig
      tags: [markets]
      summary: Update market auto-update configuration
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketAutoConfigRequest"
      responses:
        "200":
          description: Updated market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
  /markets/{id}/fetch-now:
    post:
      operationId: fetchMarketResultNow
      tags: [markets]
      summary: Manually trigger result fetch for a market
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fetch result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchResultResponse"
  /scraper/logs:
    get:
      operationId: getScraperLogs
      tags: [scraper]
      summary: Get scraper logs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: marketId
          schema:
            type: integer
      responses:
        "200":
          description: Scraper logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScraperLog"
  /scraper/status:
    get:
      operationId: getScraperStatus
      tags: [scraper]
      summary: Get scraper scheduler status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Scraper status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScraperStatus"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Market:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        isActive:
          type: boolean
        openResult:
          type: string
          nullable: true
        closeResult:
          type: string
          nullable: true
        jodiResult:
          type: string
          nullable: true
        autoUpdate:
          type: boolean
        sourceUrl:
          type: string
          nullable: true
        lastFetchedAt:
          type: string
          nullable: true
        fetchError:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - name
        - openTime
        - closeTime
        - isActive
        - autoUpdate
        - createdAt
    MarketAutoConfigRequest:
      type: object
      properties:
        autoUpdate:
          type: boolean
        sourceUrl:
          type: string
          nullable: true
      required:
        - autoUpdate
    FetchResultResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        openResult:
          type: string
          nullable: true
        closeResult:
          type: string
          nullable: true
        jodiResult:
          type: string
          nullable: true
      required:
        - success
        - message
    ScraperLog:
      type: object
      properties:
        id:
          type: integer
        marketId:
          type: integer
          nullable: true
        marketName:
          type: string
          nullable: true
        sourceUrl:
          type: string
          nullable: true
        success:
          type: boolean
        openResult:
          type: string
          nullable: true
        closeResult:
          type: string
          nullable: true
        jodiResult:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - success
        - createdAt
    ScraperStatus:
      type: object
      properties:
        isRunning:
          type: boolean
        totalMarketsWithAutoUpdate:
          type: integer
        lastRunAt:
          type: string
          nullable: true
        nextRunIn:
          type: string
      required:
        - isRunning
        - totalMarketsWithAutoUpdate
        - nextRunIn
...